+++ src/project/src/service/transfer/SameBankTransfer.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 public class SameBankTransfer implements TransferManager {
     private final AccountRepo accountRepo = InternalAccountRepo.getInstance();
 
-    public void transfer(int fromId, int toId, double amount) {
+    public void transfer(int fromId, int toId, long amount) {
         Optional<Account> from = accountRepo.accessAccount(fromId);
         Optional<Account> to = accountRepo.accessAccount(toId);
