+++ src/project/src/service/transfer/TransferManager.java
@@ -1,5 +1,5 @@
 package src.service.transfer;
 
 public interface TransferManager {
-    void transfer(int fromId, int toId, double amount);
+    void transfer(int fromId, int toId, long amount);
 }
